requirements.txt
Flask==1.1.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.4.1
Pillow==7.1.1
torch==1.5.0
torchvision==0.6.0
Werkzeug==1.0.1